requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
easydict==1.13
filelock==3.14.0
Flask==3.0.3
Flask-SocketIO==5.3.6
fsspec==2024.3.1
h11==0.14.0
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
protobuf==5.26.1
python-engineio==4.9.0
python-socketio==5.11.2
setuptools==69.5.1
simple-websocket==1.0.0
sympy==1.12
tbb==2021.12.0
tensorboardX==2.6.2.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.11.0
Werkzeug==3.0.3
wsproto==1.2.0
websocket==12.0